# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GigsHub API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export interface RegisterRequest {
  name: string;
  email: string;
  phone: string;
  /** @minLength 6 */
  password: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  user: "user",
  agent: "agent",
  admin: "admin",
} as const;

export interface User {
  id: string;
  name: string;
  email: string;
  phone: string;
  role: UserRole;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface Network {
  id: string;
  name: string;
  code: string;
  color: string;
  logoUrl?: string;
}

export type BundleType = (typeof BundleType)[keyof typeof BundleType];

export const BundleType = {
  daily: "daily",
  weekly: "weekly",
  monthly: "monthly",
  special: "special",
} as const;

export interface Bundle {
  id: string;
  networkId: string;
  networkName: string;
  name: string;
  data: string;
  validity: string;
  price: number;
  type: BundleType;
  popular: boolean;
}

export interface Service {
  id: string;
  name: string;
  description: string;
  category: string;
  price: number;
  iconUrl?: string;
}

export type OrderType = (typeof OrderType)[keyof typeof OrderType];

export const OrderType = {
  bundle: "bundle",
  afa_registration: "afa_registration",
  agent_registration: "agent_registration",
} as const;

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  pending: "pending",
  processing: "processing",
  completed: "completed",
  failed: "failed",
} as const;

export type OrderDetails = { [key: string]: unknown };

export interface Order {
  id: string;
  userId: string;
  type: OrderType;
  status: OrderStatus;
  amount: number;
  details?: OrderDetails;
  createdAt: string;
}

export type CreateOrderRequestType =
  (typeof CreateOrderRequestType)[keyof typeof CreateOrderRequestType];

export const CreateOrderRequestType = {
  bundle: "bundle",
  afa_registration: "afa_registration",
  agent_registration: "agent_registration",
} as const;

export type CreateOrderRequestDetails = { [key: string]: unknown };

export interface CreateOrderRequest {
  type: CreateOrderRequestType;
  bundleId?: string;
  phoneNumber: string;
  serviceId?: string;
  details?: CreateOrderRequestDetails;
}

export type TransactionType =
  (typeof TransactionType)[keyof typeof TransactionType];

export const TransactionType = {
  credit: "credit",
  debit: "debit",
} as const;

export interface Transaction {
  id: string;
  type: TransactionType;
  amount: number;
  description: string;
  createdAt: string;
}

export interface Wallet {
  id: string;
  userId: string;
  balance: number;
  currency: string;
  transactions?: Transaction[];
}

export type TopupRequestPaymentMethod =
  (typeof TopupRequestPaymentMethod)[keyof typeof TopupRequestPaymentMethod];

export const TopupRequestPaymentMethod = {
  momo: "momo",
  card: "card",
  bank: "bank",
} as const;

export interface TopupRequest {
  /** @minimum 1 */
  amount: number;
  paymentMethod: TopupRequestPaymentMethod;
}

export type GetBundlesParams = {
  /**
   * Filter by network provider
   */
  networkId?: string;
  /**
   * Filter by bundle type
   */
  type?: GetBundlesType;
};

export type GetBundlesType =
  (typeof GetBundlesType)[keyof typeof GetBundlesType];

export const GetBundlesType = {
  daily: "daily",
  weekly: "weekly",
  monthly: "monthly",
  special: "special",
} as const;