# lib/api-zod/src/generated/types/registerRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GigsHub API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RegisterRequest {
  name: string;
  email: string;
  phone: string;
  /** @minLength 6 */
  password: string;
}